Port staged engine to direct SDL3
Build the complete Stratagus and Wargus surface against the pinned SDL3 family while retaining the SDL2 lane, typed compatibility boundaries, reconstruction guardrails, and Release-sensitive ownership tests.


Copilot-Session: 5e7a7fae-df6a-4c41-abbb-a584d5de1c25

+++ engine/stratagus/src/include/sdl_input_adapter.h
@@ -15,7 +15,7 @@ std::vector<InputIntent> AdaptSdlTouchEvent(TouchInputState &touchInput,
                                             int width, int height,
                                             std::uint32_t timestamp, int modifiers);
 
-SdlFocusEventPolicy GetSdlFocusEventPolicy(Uint8 windowEvent,
+SdlFocusEventPolicy GetSdlFocusEventPolicy(Uint32 windowEvent,
                                             bool networkGame,
                                             bool pauseOnLeave);
